Generating inflow and outflow files in the GLM format using arrow

Generating inflow and outflow files in the GLM format using arrow

Usage

create_inflow_outflow_files(config, config_set_name, lake_directory)

Arguments

config

configuration file

config_set_name

specific name of configuration within the configuration directory

lake_directory

directory for FLARE application

Value

list with two vectors. One vector is the matrix of inflow_file_names and the other is the matrix of outflow_file_names

Details

Processes historical model data and future model files into the GLM format